requirements.txt
langgraph>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
fastmcp>=0.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.27.0
mysql-connector-python>=8.0.0
